+++ packages/gatsby-remark-autolink-headers/src/gatsby-ssr.js
@@ -13,10 +13,16 @@ exports.onRenderBody = ({ setHeadComponents }, pluginOptions) => {
   )
 
   const styles = `
-    .${className} {
-      float: left;
+    .${className}.before {
+      position: absolute;
+      top: 0;
+      left: 0;
+      transform: translateX(-100%);
       padding-right: 4px;
-      margin-left: -20px;
+    }
+    .${className}.after {
+      display: inline-block;
+      padding-left: 4px;
     }
     h1 .${className} svg,
     h2 .${className} svg,