+++ packages/gatsby-transformer-screenshot/chrome/buildChrome.sh
@@ -1,3 +1,7 @@
+#!/bin/bash
+
+# shellcheck disable=SC2129,SC1090
+
 # build headless chrome on EC2
 # https://github.com/adieuadieu/serverless-chrome/blob/master/chrome/README.md
 
@@ -5,15 +9,15 @@
 
 yum install -y git redhat-lsb python bzip2 tar pkgconfig atk-devel alsa-lib-devel bison binutils brlapi-devel bluez-libs-devel bzip2-devel cairo-devel cups-devel dbus-devel dbus-glib-devel expat-devel fontconfig-devel freetype-devel gcc-c++ GConf2-devel glib2-devel glibc.i686 gperf glib2-devel gtk2-devel gtk3-devel java-1.*.0-openjdk-devel libatomic libcap-devel libffi-devel libgcc.i686 libgnome-keyring-devel libjpeg-devel libstdc++.i686 libX11-devel libXScrnSaver-devel libXtst-devel libxkbcommon-x11-devel ncurses-compat-libs nspr-devel nss-devel pam-devel pango-devel pciutils-devel pulseaudio-libs-devel zlib.i686 httpd mod_ssl php php-cli python-psutil wdiff --enablerepo=epel
 
-cd ~
+cd ~ || exit
 git clone https://chromium.googlesource.com/chromium/tools/depot_tools.git
 echo "export PATH=$PATH:$HOME/depot_tools" >> ~/.bash_profile
 source ~/.bash_profile
 
 mkdir Chromium
-cd Chromium
+cd Chromium || exit
 fetch --no-history chromium
-cd src
+cd src || exit
 
 # use /tmp instead of /dev/shm
 # https://groups.google.com/a/chromium.org/forum/#!msg/headless-dev/qqbZVZ2IwEw/CPInd55OBgAJ
@@ -29,7 +33,7 @@ echo 'enable_nacl = false' >> out/Headless/args.gn
 gn gen out/Headless
 ninja -C out/Headless headless_shell
 
-cd out/Headless
-tar -zcvf /home/ec2-user/headless_shell.tar.gz headless_shell 
+cd out/Headless || exit
+tar -zcvf /home/ec2-user/headless_shell.tar.gz headless_shell
 
 # scp ec2-user@xxx.amazonaws.com:~/headless_shell.tar.gz .