+++ packages/gatsby/cache-dir/production-app.js
@@ -26,7 +26,6 @@ loader.setApiRunner(apiRunner)
 window.asyncRequires = asyncRequires
 window.___emitter = emitter
 window.___loader = publicLoader
-window.___webpackCompilationHash = window.webpackCompilationHash
 
 navigationInit()
 
@@ -73,12 +72,14 @@ apiRunnerAsync(`onClientEntry`).then(() => {
                   id="gatsby-focus-wrapper"
                 >
                   <RouteHandler
-                    path={encodeURI(
+                    path={
                       pageResources.page.path === `/404.html`
                         ? stripPrefix(location.pathname, __BASE_PATH__)
-                        : pageResources.page.matchPath ||
-                            pageResources.page.path
-                    )}
+                        : encodeURI(
+                            pageResources.page.matchPath ||
+                              pageResources.page.path
+                          )
+                    }
                     {...this.props}
                     location={location}
                     pageResources={pageResources}
@@ -123,6 +124,9 @@ apiRunnerAsync(`onClientEntry`).then(() => {
         `page resources for ${browserLoc.pathname} not found. Not rendering React`
       )
     }
+
+    window.___webpackCompilationHash = page.page.webpackCompilationHash
+
     const Root = () => (
       <Location>
         {locationContext => <LocationHandler {...locationContext} />}