+++ starters/blog/src/templates/blog-post.js
@@ -87,7 +87,6 @@ export const pageQuery = graphql`
     site {
       siteMetadata {
         title
-        author
       }
     }
     markdownRemark(fields: { slug: { eq: $slug } }) {