+++ www/src/components/layout.js
@@ -3,12 +3,17 @@ import { jsx } from "theme-ui"
 import React from "react"
 import { navigate, PageRenderer } from "gatsby"
 import mousetrap from "mousetrap"
-import Modal from "react-modal"
 import MdClose from "react-icons/lib/md/close"
+
 import { Global } from "@emotion/core"
 
 import { globalStyles } from "../utils/styles/global"
-import { colors, space, zIndices } from "../gatsby-plugin-theme-ui"
+import {
+  colors,
+  space,
+  zIndices,
+  mediaQueries,
+} from "../gatsby-plugin-theme-ui"
 import { breakpointGutter } from "../utils/styles"
 import Banner from "../components/banner"
 import withColorMode from "../components/with-color-mode"
@@ -18,6 +23,7 @@ import PageWithSidebar from "../components/page-with-sidebar"
 import SiteMetadata from "../components/site-metadata"
 import SkipNavLink from "../components/skip-nav-link"
 import "../assets/fonts/futura"
+import LazyModal from "./lazy-modal"
 
 let windowWidth
 
@@ -32,8 +38,6 @@ class DefaultLayout extends React.Component {
   }
 
   componentDidMount() {
-    Modal.setAppElement(`#___gatsby`)
-
     if (this.props.isModal && window.innerWidth > 750) {
       mousetrap.bind(`left`, this.props.modalPrevious)
       mousetrap.bind(`right`, this.props.modalNext)
@@ -73,7 +77,7 @@ class DefaultLayout extends React.Component {
           <PageRenderer
             location={{ pathname: this.props.modalBackgroundPath }}
           />
-          <Modal
+          <LazyModal
             isOpen={true}
             style={{
               content: {
@@ -86,7 +90,7 @@ class DefaultLayout extends React.Component {
                 padding: `${space[8]} 0`,
                 right: `inherit`,
                 top: `inherit`,
-                width: `750px`,
+                maxWidth: `1050px`,
               },
               overlay: {
                 backgroundColor: isDark
@@ -108,40 +112,54 @@ class DefaultLayout extends React.Component {
           >
             <div
               sx={{
-                bg: `card.background`,
-                borderRadius: 2,
-                boxShadow: `dialog`,
-                position: `relative`,
+                display: `flex`,
+                flexWrap: `wrap`,
+                justifyContent: `space-between`,
+                [mediaQueries.md]: {
+                  flexWrap: `nowrap`,
+                },
               }}
             >
-              <button
-                onClick={this.handleCloseModal}
+              <div
                 sx={{
                   bg: `card.background`,
-                  border: 0,
-                  borderRadius: 6,
-                  color: `textMuted`,
-                  cursor: `pointer`,
-                  fontSize: 4,
-                  height: 40,
-                  left: `auto`,
-                  position: `absolute`,
-                  right: t => t.space[7],
-                  top: t => t.space[8],
-                  width: 40,
-                  "&:hover": {
-                    bg: `ui.hover`,
-                    color: `gatsby`,
-                  },
+                  borderRadius: 2,
+                  boxShadow: `dialog`,
+                  position: `relative`,
+                  alignItems: `center`,
+                  order: 1,
+                  width: `100%`,
                 }}
               >
-                <MdClose />
-              </button>
-              {this.props.children}
+                <button
+                  onClick={this.handleCloseModal}
+                  sx={{
+                    bg: `card.background`,
+                    border: 0,
+                    borderRadius: 6,
+                    color: `textMuted`,
+                    cursor: `pointer`,
+                    fontSize: 4,
+                    height: 40,
+                    left: `auto`,
+                    position: `absolute`,
+                    right: t => t.space[7],
+                    top: t => t.space[8],
+                    width: 40,
+                    "&:hover": {
+                      bg: `ui.hover`,
+                      color: `gatsby`,
+                    },
+                  }}
+                >
+                  <MdClose />
+                </button>
+                {this.props.children}
+              </div>
               {this.props.modalPreviousLink}
               {this.props.modalNextLink}
             </div>
-          </Modal>
+          </LazyModal>
         </>
       )
     }