+++ packages/gatsby-graphiql-explorer/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.31](https://github.com/gatsbyjs/gatsby/compare/gatsby-graphiql-explorer@0.2.29...gatsby-graphiql-explorer@0.2.31) (2019-12-10)
+
+**Note:** Version bump only for package gatsby-graphiql-explorer
+
+## [0.2.30](https://github.com/gatsbyjs/gatsby/compare/gatsby-graphiql-explorer@0.2.29...gatsby-graphiql-explorer@0.2.30) (2019-12-10)
+
+**Note:** Version bump only for package gatsby-graphiql-explorer
+
+## [0.2.29](https://github.com/gatsbyjs/gatsby/compare/gatsby-graphiql-explorer@0.2.28...gatsby-graphiql-explorer@0.2.29) (2019-11-26)
+
+**Note:** Version bump only for package gatsby-graphiql-explorer
+
+## [0.2.28](https://github.com/gatsbyjs/gatsby/compare/gatsby-graphiql-explorer@0.2.27...gatsby-graphiql-explorer@0.2.28) (2019-11-15)
+
+**Note:** Version bump only for package gatsby-graphiql-explorer
+
+## [0.2.27](https://github.com/gatsbyjs/gatsby/compare/gatsby-graphiql-explorer@0.2.26...gatsby-graphiql-explorer@0.2.27) (2019-11-10)
+
+**Note:** Version bump only for package gatsby-graphiql-explorer
+
+## [0.2.26](https://github.com/gatsbyjs/gatsby/compare/gatsby-graphiql-explorer@0.2.25...gatsby-graphiql-explorer@0.2.26) (2019-10-28)
+
+### Bug Fixes
+
+- update minor updates in packages ([#18875](https://github.com/gatsbyjs/gatsby/issues/18875)) ([b692879](https://github.com/gatsbyjs/gatsby/commit/b692879))
+
 ## [0.2.25](https://github.com/gatsbyjs/gatsby/compare/gatsby-graphiql-explorer@0.2.24...gatsby-graphiql-explorer@0.2.25) (2019-10-14)
 
 **Note:** Version bump only for package gatsby-graphiql-explorer